requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.2
bcrypt>=4.1.2
SQLAlchemy>=2.0.29
pydantic>=2.7.1
openai>=1.30.0
tiktoken>=0.7.0
rapidfuzz>=3.9.6
